One black-hole tunnel, not one per test class
The connect-failure trigger and the retirement test were written on separate
branches a day apart and each grew its own copy of the same fixture, along with
the same bound-then-released port helper. Now shared, with the reasoning in one
place: why removing a node from the fake does not model a node going away, and
why refused beats dropped (a dropped SYN exercises connect timeouts instead,
which is a different failure mode and a slower test).
